+++ src/hooks/queries/index.ts
@@ -1,11 +1,13 @@
 import useFetchArticle from "./useFetchArticle";
 import useFetchArticleDetail from "./useFetchArticleDetail";
 import useCreateArticle from "./useCreateArticle";
-import useDeleteArticleDetail from "./useDeleteArticleDetail";
+import useDeleteArticleDetail from "./useDeleteArticle";
+import useUpdateArticle from "./useUpdateArticle";
 
 export {
   useFetchArticle,
   useFetchArticleDetail,
   useCreateArticle,
   useDeleteArticleDetail,
+  useUpdateArticle,
 };
